refactor: merge webpack api parser plugins
