+++ crates/rspack_plugin_javascript/src/parser_plugin/mod.rs
@@ -10,7 +10,6 @@ mod esm_detection_parser_plugin;
 mod esm_export_dependency_parser_plugin;
 mod esm_import_dependency_parser_plugin;
 mod esm_top_level_this_plugin;
-mod exports_info_api_plugin;
 mod import_meta_context_dependency_parser_plugin;
 mod import_meta_plugin;
 mod import_parser_plugin;
@@ -18,7 +17,6 @@ mod import_phase;
 mod initialize_evaluating;
 mod inline_const;
 pub(crate) mod inner_graph;
-mod is_included_plugin;
 mod javascript_meta_info_plugin;
 pub mod node_stuff_plugin;
 mod override_strict_plugin;
@@ -54,14 +52,12 @@ pub(crate) use self::{
   esm_export_dependency_parser_plugin::ESMExportDependencyParserPlugin,
   esm_import_dependency_parser_plugin::ESMImportDependencyParserPlugin,
   esm_top_level_this_plugin::ESMTopLevelThisParserPlugin,
-  exports_info_api_plugin::ExportsInfoApiPlugin,
   import_meta_context_dependency_parser_plugin::ImportMetaContextDependencyParserPlugin,
   import_meta_plugin::{ImportMetaDisabledPlugin, ImportMetaPlugin},
   import_parser_plugin::{ImportParserPlugin, ImportsReferencesState},
   initialize_evaluating::InitializeEvaluating,
   inline_const::{ConstValue, ConstValuePlugin},
   inner_graph::{connection_active_used_by_exports, plugin::*, runtime_condition_used_by_exports},
-  is_included_plugin::IsIncludedPlugin,
   javascript_meta_info_plugin::JavascriptMetaInfoPlugin,
   node_stuff_plugin::NodeStuffPlugin,
   override_strict_plugin::OverrideStrictPlugin,
